First integration tests to check the correct flow of the contact controller, services, database connections and CRUD operation in the database.

+++ src/main/java/fabio/dev/Portfolio/Services/ContactService.java
@@ -1,16 +1,21 @@
-package fabio.dev.Portfolio;
+package fabio.dev.Portfolio.Services;
 
 import fabio.dev.Portfolio.DTOs.ContactDTO;
 import fabio.dev.Portfolio.DTOs.ContactUpdateDTO;
+import fabio.dev.Portfolio.Exceptions.NoEntityException;
 import fabio.dev.Portfolio.Models.Contact;
 import fabio.dev.Portfolio.Repositorys.ContactRepository;
 import jakarta.transaction.Transactional;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
-public class ContactService {
+public class ContactService{
+
+    private Logger logger = LoggerFactory.getLogger(ContactService.class);
 
     private ContactRepository contactRepository;
 
@@ -20,35 +25,55 @@ public class ContactService {
 
     @Transactional
     public Contact saveContact(ContactDTO dto){
+
+        logger.info("saving new contact");
+
         Contact contact = new Contact();
         contact.setName(dto.getName());
         contact.setEmail(dto.getEmail());
         contact.setMessage(dto.getMessage());
 
+        logger.info("saved contact with id {} successfuly", contact.getId());
         return contactRepository.save(contact);
     }
 
     public List<Contact> findAllContacts(){
+        logger.info("founded all contacts successfully");
         return contactRepository.findAll();
     }
 
     @Transactional
-    public Contact updateContact(Integer id,ContactUpdateDTO dto){
+    public Contact updateContact(Integer id, ContactUpdateDTO dto){
 
-        Contact contact =  contactRepository.findById(id).orElseThrow(() -> new RuntimeException(""));
+        logger.info("updating contact with id {}", id);
+
+        if( id == null || id <= 0){
+            throw new NoEntityException("Invalid id");
+        }
+
+        Contact contact = contactRepository.findById(id).orElseThrow(() ->  new NoEntityException("Contact",id));
 
         if (dto.name() != null){contact.setName(dto.name());}
         if (dto.email() != null){contact.setEmail(dto.email());}
         if (dto.message() != null){contact.setMessage(dto.message());}
+
+        logger.info("updated contact with id {} successfully", id);
 
         return contactRepository.save(contact);
     }
 
     @Transactional
     public void deleteContact(Integer id){
-        if(!contactRepository.existsById(id)){
-            throw new Error("");
+        logger.info("deleting contact with id {}", id);
+
+        if( id == null || id <= 0){
+            throw new NoEntityException("Contact",id);
         }
+
+        if(!contactRepository.existsById(id)){
+            throw new NoEntityException("Contact",id);
+        }
+        logger.info("deleted contact with id {} successfully", id);
         contactRepository.deleteById(id);
     }
